+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.6.1] - 2024-04-15
+### Fixed
+ARGUS-43192
+- `ServiceV1HttpClient` throws `ServiceTimeoutException` on connect timeout.
+- Added configuration option `ServiceV1HttpClient.setConnectionTimeoutSeconds(long)` to configure non-default connection timeout. 
+Defaults to 3 minutes (HttpClient default). 
+
 ## [0.6.0] - 2024-03-20
 ### Changed
 ARGUS-39520